+++ README.md
@@ -4,6 +4,8 @@
 |:----:|:---:|
 | [![devtools-launchpad-version]][devtools-launchpad-pkg] | [devtools-launchpad] |
 | [![devtools-config-version]][devtools-config-pkg] | [devtools-config] |
+| [![devtools-components-version]][devtools-components-pkg] | [devtools-components] |
+| [![devtools-connection-version]][devtools-connection-pkg] | [devtools-connection] |
 | [![devtools-modules-version]][devtools-modules-pkg] | [devtools-modules] |
 | [![devtools-source-editor-version]][devtools-source-editor-pkg] | [devtools-source-editor] |
 | [![devtools-source-map-version]][devtools-source-map-pkg] | [devtools-source-map] |
@@ -27,9 +29,17 @@
 [devtools-config-pkg]:https://npmjs.org/package/devtools-config
 [devtools-config]:./packages/devtools-config/#readme
 
+[devtools-components-version]:https://img.shields.io/npm/v/devtools-components.svg
+[devtools-components-pkg]:https://npmjs.org/package/devtools-components
+[devtools-components]:./packages/devtools-components/#readme
+
 [devtools-modules-version]:https://img.shields.io/npm/v/devtools-modules.svg
 [devtools-modules-pkg]:https://npmjs.org/package/devtools-modules
 [devtools-modules]:./packages/devtools-modules/#readme
+
+[devtools-connection-version]:https://img.shields.io/npm/v/devtools-connection.svg
+[devtools-connection-pkg]:https://npmjs.org/package/devtools-connection
+[devtools-connection]:./packages/devtools-connection/#readme
 
 [devtools-source-editor-version]:https://img.shields.io/npm/v/devtools-source-editor.svg
 [devtools-source-editor-pkg]:https://npmjs.org/package/devtools-source-editor
