+++ packages/devtools-components/README.md
@@ -1,0 +1,5 @@
+## Devtools Components
+
+[![Npm version](https://img.shields.io/npm/v/devtools-components.svg)](https://npmjs.org/package/devtools-components)
+
+Devtools shared Components
