+++ packages/devtools-components/index.js
@@ -1,0 +1,5 @@
+const Tree = require("./src/tree");
+
+module.exports = {
+  Tree
+};
