+++ packages/devtools-components/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "devtools-modules",
-  "version": "0.0.26",
-  "description": "DevTools Modules from M-C",
+  "name": "devtools-components",
+  "version": "0.0.1",
+  "description": "DevTools HTML Components",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "author": "",
+  "author": "Jason Laster",
   "license": "ISC",
   "dependencies": {
   }
