+++ packages/devtools-connection/README.md
@@ -1,0 +1,5 @@
+## Devtools Connection
+
+[![Npm version](https://img.shields.io/npm/v/devtools-connection.svg)](https://npmjs.org/package/devtools-connection)
+
+Host firefox devtools connection logic
