+++ packages/devtools-launchpad/package.json
@@ -1,6 +1,6 @@
 {
   "name": "devtools-launchpad",
-  "version": "0.0.68",
+  "version": "0.0.71",
   "license": "MPL-2.0",
   "description": "The Launchpad makes it easy to build a developer tool for Firefox, Chrome, and Node.",
   "repository": {
@@ -48,13 +48,14 @@
     "babel-register": "^6.18.0",
     "body-parser": "^1.15.2",
     "check-node-version": "^1.1.2",
-    "classnames": "^2.2.5",
     "chrome-remote-interface": "0.17.0",
+    "classnames": "^2.2.5",
     "co": "=4.6.0",
     "css-loader": "^0.26.1",
     "devtools-config": "^0.0.12",
-    "devtools-modules": "^0.0.25",
     "devtools-connection": "^0.0.2",
+    "devtools-modules": "^0.0.26",
+    "devtools-utils": "^0.0.2",
     "eslint": "^3.12.0",
     "eslint-plugin-babel": "^3.3.0",
     "eslint-plugin-flowtype": "^2.20.0",
@@ -85,6 +86,7 @@
     "react-redux": "4.4.5",
     "redux": "3.5.2",
     "selenium-webdriver": "^3.0.1",
+    "devtools-sprintf-js": "^1.0.3",
     "style-loader": "^0.13.1",
     "svg-inline-loader": "^0.7.1",
     "webpack": "1.14.0",
