--- packages/devtools-modules/shared/clipboard.js
@@ -1,16 +1,0 @@
-/**
- * Clipboard function taken from
- * https://dxr.mozilla.org/mozilla-central/source/devtools/shared/platform/content/clipboard.js
- */
-function copyString(string) {
-  let doCopy = function (e) {
-    e.clipboardData.setData("text/plain", string);
-    e.preventDefault();
-  };
-
-  document.addEventListener("copy", doCopy);
-  document.execCommand("copy", false, null);
-  document.removeEventListener("copy", doCopy);
-}
-
-module.exports = { copyString };
