+++ packages/devtools-modules/src/menu/index.js
@@ -1,12 +1,4 @@
-/* -*- indent-tabs-mode: nil; js-indent-level: 2 -*- */
-/* vim: set ft=javascript ts=2 et sw=2 tw=80: */
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-"use strict";
-
-const EventEmitter = require("../../shared/event-emitter");
+const EventEmitter = require("../utils/event-emitter");
 
 /**
  * A partial implementation of the Menu API provided by electron:
