+++ packages/devtools-modules/src/menu/menu-item.js
@@ -1,11 +1,3 @@
-/* -*- indent-tabs-mode: nil; js-indent-level: 2 -*- */
-/* vim: set ft=javascript ts=2 et sw=2 tw=80: */
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-"use strict";
-
 /**
  * A partial implementation of the MenuItem API provided by electron:
  * https://github.com/electron/electron/blob/master/docs/api/menu-item.md.
