+++ packages/devtools-modules/src/prefs.js
@@ -2,8 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-const Services = require("./shim/Services");
-const EventEmitter = require("../../shared/event-emitter");
+const Services = require("./Services");
+const EventEmitter = require("./utils/event-emitter");
 
 /**
  * Shortcuts for lazily accessing and setting various preferences.
