+++ packages/devtools-modules/src/utils/event-emitter.js
