+++ packages/devtools-modules/src/utils/promise.js
